User Server, Homework 5 version 1.4

+++ src/events/logger.ts
@@ -1,18 +1,26 @@
 import {EventEmitter} from "node:events";
 import fs from 'fs';
-import path from "node:path";
 
 class Logger extends EventEmitter {
     private filePath: string = "";
     private logArray: Array<{ date: string, message: string }> = [];
 
     createLogFile() {
-        const start = new Date().toISOString();
-
-        this.filePath = './logs/log_' + start.replace(/:/g, '-') + '.txt';
-        fs.writeFile(this.filePath, "=============Server log file=============\n", err => {
-            if (err) myLogger.emit('saved', "log file cannot be created")
-        })
+            const start = new Date().toISOString();
+            const directory = './logs/';
+            this.filePath = './logs/log_' + start.replace(/:/g, '-') + '.txt';
+            fs.mkdir(directory, (err) => {
+                if (err && err.code !== 'EEXIST') this.save("log directory not created")
+                else {
+                    fs.writeFile(this.filePath, "=============Server log file=============\n", err => {
+                        if (err) this.save("log file cannot be created")
+                        else {
+                            this.saveToFile("session starts");
+                            this.saveToFile("server successfully started");
+                        }
+                    })
+                }
+            })
     }
 
     addLogToArray(message: string) {
@@ -27,20 +35,12 @@ class Logger extends EventEmitter {
         this.emit('logged', message);
     }
 
-    saveToFile(message: string) {
-        this.emit('saveToFile', this.filePath, message);
+    save(message: string) {
+        this.emit('saved', message);
     }
 
-    async clearLogDir() {
-        const directory = './logs/';
-        fs.readdir(directory, (err, files) => {
-            if (err) console.log("can't clean log directory")
-            else {
-                files.forEach(file => {
-                    if (file.includes('log')) fs.unlink(path.join(directory, file), (err) => {});
-                })
-            }
-        })
+    saveToFile(message: string) {
+        this.emit('saveToFile', this.filePath, message);
     }
 }
 
@@ -50,15 +50,20 @@ myLogger.on('logged', (message: string) => {
     console.log(new Date().toISOString(), message)
 });
 
+myLogger.on('saved', (message: string) => {
+    console.log(new Date().toISOString(), message)
+    myLogger.addLogToArray(message);
+});
+
 myLogger.on('saveToFile', (filePath: string, message: string) => {
     myLogger.addLogToArray(message);
     console.log(new Date().toISOString(), message)
 
     fs.access(filePath, fs.constants.F_OK, (err: any) => {
-        if (err) myLogger.emit('saved', "error - log file not found");
+        if (err) myLogger.save("error - log file not found");
         else {
             fs.appendFile(filePath, new Date().toISOString() + ' ' + message + '\n', (err: any) => {
-                if (err) myLogger.emit('saved', "error - logs can't be saved to file");
+                if (err) myLogger.save("error - logs can't be saved to file");
             })
         }
     });
